travel_app_backend/requirements.txt
Flask>=2.0
google-cloud-firestore>=2.5 # Firebase dependency
# Add other Google Cloud client libraries as needed:
google-cloud-translate>=2.0
google-cloud-vision>=2.5
# google-cloud-speech # If doing speech-to-text
googlemaps # For Google Maps APIs
requests # For making calls to other third-party APIs
# Add any other dependencies (e.g., for ML models if built in Python)
gunicorn # Recommended for running Flask in production environments (like Cloud Run)
python-dotenv # To load environment variables from a .env file for local dev